Move workout view to exercise package
