+++ app/src/main/java/com/example/workout_companion/view/ExerciseOverview.kt
@@ -3,6 +3,7 @@ package com.example.workout_companion.view
 import androidx.compose.material.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
+import com.example.workout_companion.view.exercise.WorkoutsView
 import com.example.workout_companion.view.inputfields.TopNavigation
 
 @Composable
